Fix build problem building ipl/cfuncs/ipl with musl library
When building with the musl library (an alternative to glibc), fpoll.c
fails due to the sensitivity of the internals of the FILE structure.

The musl library provides a function (__freadahead) that performs the
necessary test to see if data is available, however musl does not provide
a method to determine if the musl library is in use.

In order to build with the musl library, define the macro _MUSL via the
CFLAGS variable in Makedefs (e.g. CFLAGS=$(CFLAGS) -D_MUSL)