Ensure obj dataype in data.r match obj definition
When building with lto (Link Time Optimization) flags, the compiler
discovered an inconsistency in the datatypes for the structures:
    BbScan, Bescan, Bfield, Blimit, and Bllist.

    [gcc-15] data.r: error: type of Bllist does not match original
        declaration [-Werror=lto-type-mismatch]

In rmacros.h, the macro OpBlock produces a structure definition of:

    struct b_iproc B{name} ...

In grttin.h the macro LibDcl is uses OpBlock to define structures.

The definitions for the structures BbScan, Bescan, Bfield, Blimit, and
Bllist are generated in imisc.r and omisc.r via LibDcl's or OpBlock's.

However in data.r, the datatype for these structures are declared as
struct b_proc.

To fix the lto problem, change the data type of the externs and the
datatype of the opblks array to be b_iproc instead of b_proc so they
properly match the datatypes in imisc and omisc.

Note this problem was discovered via gentoo's QA processing
(see https://bugs.gentoo.org/940325)